chore: rename app global folder to shared
