+++ src/components/app/shared/footer/AppFooter.tsx
