+++ src/components/app/shared/player/mini-player/MiniPlayer.tsx
