+++ src/components/app/shared/player/shared/device-menu/DeviceIcon.tsx
