--- src/components/app/views/login/PostAuthFail.tsx
@@ -1,3 +1,0 @@
-export const PostAuthFail = () => {
-  return <p>post auth fail</p>;
-};
