+++ src/components/web/pages/404/404.tsx
@@ -24,10 +24,10 @@ export const _404 = () => {
         <p>Otherwise we hope these options are somewhat helpful to you:</p>
         <ul>
           <li>
-            <Link to={routes.web.root}>Home</Link>
+            <Link to={routes.root}>Home</Link>
           </li>
           <li>
-            <Link to={routes.web.documentation}>Documentation</Link>
+            <Link to={routes.documentation}>Documentation</Link>
           </li>
         </ul>
       </main>
