+++ src/utils/public-env.ts
@@ -1,0 +1,6 @@
+enum env {
+  CLIENT_ID = '5ca64c0a829949428154075795560d0d',
+  REDIRECT_URL = 'https://localhost:5173/auth',
+}
+
+export default env;
